+++ src/main/java/in/koreatech/koin/domain/shop/dto/CreateMenuRequest.java
@@ -14,7 +14,7 @@ import jakarta.validation.constraints.Size;
 @JsonNaming(value = SnakeCaseStrategy.class)
 public record CreateMenuRequest(
     @Schema(example = "0", description = "선택된 카테고리 고유 id 리스트")
-    @NotNull List<Long> categoryIds,
+    @NotNull List<Integer> categoryIds,
 
     @Schema(example = "저희 가게의 대표 메뉴 짜장면입니다.", description = "메뉴 구성 설명")
     @Size(max = 80) String description,
@@ -34,7 +34,8 @@ public record CreateMenuRequest(
     @Schema(description = "단일 메뉴일때의 가격")
     Integer singlePrice
 ) {
-    public Menu toEntity(Long shopId) {
+
+    public Menu toEntity(Integer shopId) {
         return Menu.builder()
             .name(name)
             .shopId(shopId)
