+++ src/main/java/in/koreatech/koin/domain/shop/repository/ShopRepository.java
@@ -8,22 +8,22 @@ import org.springframework.data.repository.Repository;
 import in.koreatech.koin.domain.shop.exception.ShopNotFoundException;
 import in.koreatech.koin.domain.shop.model.Shop;
 
-public interface ShopRepository extends Repository<Shop, Long> {
+public interface ShopRepository extends Repository<Shop, Integer> {
 
     Shop save(Shop shop);
 
-    List<Shop> findAllByOwnerId(Long ownerId);
+    List<Shop> findAllByOwnerId(Integer ownerId);
 
-    Optional<Shop> findById(Long shopId);
+    Optional<Shop> findById(Integer shopId);
 
-    Optional<Shop> findByOwnerId(Long ownerId);
+    Optional<Shop> findByOwnerId(Integer ownerId);
 
-    default Shop getById(Long shopId) {
+    default Shop getById(Integer shopId) {
         return findById(shopId)
             .orElseThrow(() -> ShopNotFoundException.withDetail("shopId: " + shopId));
     }
 
-    default Shop getByOwnerId(Long ownerId) {
+    default Shop getByOwnerId(Integer ownerId) {
         return findByOwnerId(ownerId)
             .orElseThrow(() -> ShopNotFoundException.withDetail("ownerId: " + ownerId));
     }
