+++ src/main/java/in/koreatech/koin/global/auth/AuthArgumentResolver.java
@@ -35,7 +35,7 @@ public class AuthArgumentResolver implements HandlerMethodArgumentResolver {
 
     @Override
     public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
-                                  NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
+        NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
 
         Auth authAt = parameter.getParameterAnnotation(Auth.class);
         requireNonNull(authAt);
@@ -43,7 +43,7 @@ public class AuthArgumentResolver implements HandlerMethodArgumentResolver {
         if (authContext.isAnonymous() && authAt.anonymous()) {
             return null;
         }
-        Long userId = authContext.getUserId();
+        Integer userId = authContext.getUserId();
         User user = userRepository.getById(userId);
 
         if (permitStatus.contains(user.getUserType())) {
@@ -58,7 +58,7 @@ public class AuthArgumentResolver implements HandlerMethodArgumentResolver {
             }
             return user.getId();
         }
-        HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
+        HttpServletRequest request = (HttpServletRequest)webRequest.getNativeRequest();
         throw AuthorizationException.withDetail("header: " + request);
     }
 }
